pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "BeamClean"
version = "0.1.0"
description = "Library for BeamClean"
authors = ["Kaan Kale"]
license = "MIT"
readme = "README.md"
package-mode = false
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
accelerate = "^1.3.0"
datasets = "*"
freia = "^0.2"
huggingface_hub = "*"
hydra-core = "^1.3.2"
ipywidgets = "*"
jinja2 = "3.1.6"
matplotlib = "^3.10.0"
omegaconf = "^2.3.0"
psutil = "5.9.0"
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
seaborn = "^0.13.2"
scikit-learn = "^1.6.0"
torch = "2.5.1"
transformers = "^4.48.0"
wandb = "^0.19.1"
zuko = "^1.3.1"
hypothesis = "^6.130.5"
evaluate = "^0.4.3"
nltk = "^3.9.1"
rouge-score = "^0.1.2"
lm-eval = "^0.4.8"
jaxtyping = "^0.3.2"
dspy = "^2.6.23"



[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
ipykernel = "^6.29.5"
